pyproject.toml
[tool.poetry]
name = "garpy"
version = "0.1.0"
description = "Python client for downloading activities from Garmin Connect"
authors = ["Felipe Aguirre Martinez <felipeam86@gmail.com>"]
license = "Apache-2.0"
readme = "README.md"


[tool.poetry.dependencies]
python = "^3.7"
requests = "^2.22"
attrs = "^19.1"

[tool.poetry.dev-dependencies]
pytest = "^3.0"

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
